requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-cloud-firestore
google-cloud-storage
structlog
python-dotenv
pytest
mock-firestore
python-dateutil
python-multipart
